[SDK] Rename verifyPayment() to processPayment() for x402 payments
