+++ .changeset/some-moons-burn.md
@@ -2,4 +2,4 @@
 "thirdweb": minor
 ---
 
-Accept arbitrary chain ids for x402 payments with new verifyPayment() backend utility
+Accept arbitrary chain ids for x402 payments with new settlePayment() and verifyPayment() backend utility functions
